+++ src/Presentation/Nop.Web/Areas/Admin/Models/Catalog/CrossSellProductModel.cs
@@ -10,7 +10,7 @@ public partial record CrossSellProductModel : BaseNopEntityModel
 {
     #region Properties
 
-    public int ProductId2 { get; set; }
+    public long ProductId2 { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.CrossSells.Fields.Product")]
     public string Product2Name { get; set; }
