+++ src/Presentation/Nop.Web/Areas/Admin/Models/Catalog/ProductSearchModel.cs
@@ -32,28 +32,28 @@ public partial record ProductSearchModel : BaseSearchModel
     public string SearchProductName { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchCategory")]
-    public int SearchCategoryId { get; set; }
+    public long SearchCategoryId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchIncludeSubCategories")]
     public bool SearchIncludeSubCategories { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchManufacturer")]
-    public int SearchManufacturerId { get; set; }
+    public long SearchManufacturerId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchStore")]
-    public int SearchStoreId { get; set; }
+    public long SearchStoreId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchVendor")]
-    public int SearchVendorId { get; set; }
+    public long SearchVendorId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchWarehouse")]
-    public int SearchWarehouseId { get; set; }
+    public long SearchWarehouseId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchProductType")]
-    public int SearchProductTypeId { get; set; }
+    public long SearchProductTypeId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.SearchPublished")]
-    public int SearchPublishedId { get; set; }
+    public long SearchPublishedId { get; set; }
 
     [NopResourceDisplayName("Admin.Catalog.Products.List.GoDirectlyToSku")]
     public string GoDirectlyToSku { get; set; }
