+++ src/Presentation/Nop.Web/Areas/Admin/Models/Common/UrlRecordSearchModel.cs
@@ -25,10 +25,10 @@ public partial record UrlRecordSearchModel : BaseSearchModel
     public string SeName { get; set; }
 
     [NopResourceDisplayName("Admin.System.SeNames.List.Language")]
-    public int LanguageId { get; set; }
+    public long LanguageId { get; set; }
 
     [NopResourceDisplayName("Admin.System.SeNames.List.IsActive")]
-    public int IsActiveId { get; set; }
+    public long IsActiveId { get; set; }
 
     public IList<SelectListItem> AvailableLanguages { get; set; }
