+++ src/Presentation/Nop.Web/Areas/Admin/Models/Customers/AddProductToCustomerRoleModel.cs
@@ -9,7 +9,7 @@ public partial record AddProductToCustomerRoleModel : BaseNopEntityModel
 {
     #region Properties
 
-    public int AssociatedToProductId { get; set; }
+    public long AssociatedToProductId { get; set; }
 
     #endregion
 }
