+++ src/Presentation/Nop.Web/Areas/Admin/Models/Discounts/DiscountRequirementRuleModel.cs
@@ -19,7 +19,7 @@ public partial record DiscountRequirementRuleModel : BaseNopModel
 
     #region Properties
 
-    public int DiscountRequirementId { get; set; }
+    public long DiscountRequirementId { get; set; }
 
     public string RuleName { get; set; }
 
@@ -27,7 +27,7 @@ public partial record DiscountRequirementRuleModel : BaseNopModel
 
     public string InteractionType { get; set; }
 
-    public int? ParentId { get; set; }
+    public long? ParentId { get; set; }
 
     public SelectList AvailableInteractionTypes { get; set; }
